all: Make the block number in EthereumBlockPointer a BlockNumber
We can't handle u64 block numbers - any interaction with the store with a
block number that does not fit into an i32 will blow up. Rather than do the
conversion of u64 to BlockNumber adhoc and in many places, force the code
to do those conversions when we take in a block number from an external
source.

The code could be tightened up further by using a newtype for BlockNumber,
but that is left for another change.

+++ core/src/subgraph/registrar.rs
@@ -472,11 +472,7 @@ fn resolve_subgraph_chain_blocks(
                     let base_block = base.block;
                     Box::new(
                         ethereum_adapter
-                            .block_pointer_from_number(
-                                &logger1,
-                                chain_store1.clone(),
-                                base.block as u64,
-                            )
+                            .block_pointer_from_number(&logger1, chain_store1.clone(), base.block)
                             .map(|ptr| Some((base.base, ptr)))
                             .map_err(move |_| {
                                 SubgraphRegistrarError::ManifestValidationError(vec![
